server/requirements.txt
aniso8601==8.0.0
APScheduler==3.6.3
cachetools==4.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Dance==3.0.0
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-RESTful==0.3.8
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
google-api-core==1.17.0
google-api-python-client==1.8.0
google-auth==1.14.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.2
idna==2.8
iexfinance==0.4.3
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.3
oauthlib==3.0.1
pandas==1.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyJWT==1.4.2
pyOpenSSL==19.0.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-http-client==3.2.7
pytz==2019.3
redis==3.4.1
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
sendgrid==6.3.1
six==1.13.0
SQLAlchemy==1.3.12
SQLAlchemy-Utils==0.36.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.24.3
URLObject==2.4.3
Werkzeug==0.16.0
